% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_whole_number}
\alias{.is_whole_number}
\title{Check Whether a Value is a Whole Number}
\usage{
.is_whole_number(x)
}
\arguments{
\item{x}{Value to test.}
}
\value{
Logical scalar.
}
\description{
Check Whether a Value is a Whole Number
}
\keyword{internal}
